Refactor profile module into SRP submodules
